requirements.txt
discord.py>=2.4.0
humanfriendly==10.0
psutil==6.0.0
PyYAML==6.0.2
Requests==2.32.3
